Simplification of the DynamoDb code.
