+++ Cargo.toml
@@ -33,11 +33,11 @@ async-graphql-axum = "5.0.10"
 async-lock = "2.8.0"
 async-stream = "0.3.5"
 async-trait = "0.1.74"
-aws-config = "0.14.0"
-aws-sdk-dynamodb = "0.14.0"
-aws-sdk-s3 = "0.14.0"
-aws-smithy-http = "0.44.0"
-aws-types = "0.14.0"
+aws-config = "0.56.1"
+aws-sdk-dynamodb = "0.34.0"
+aws-sdk-s3 = "0.34.0"
+aws-smithy-http = "0.56.1"
+aws-types = "0.56.1"
 scylla = "0.8.2"
 axum = "0.6.20"
 bcs = "0.1.6"
